requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
torch==2.0.1
transformers==4.36.2
colorama==0.4.6
numpy==1.24.3
